Insert module agenda into main

+++ src/main/java/imdl/scalator/service/InstrumentoService.java
@@ -2,6 +2,7 @@ package imdl.scalator.service;
 
 import imdl.scalator.domain.Instrumento;
 import imdl.scalator.domain.exception.EntityNotFoundException;
+import imdl.scalator.domain.exception.RogueException;
 import imdl.scalator.domain.input.InstrumentoInput;
 import imdl.scalator.persistence.InstrumentoRepository;
 import imdl.scalator.service.mapper.InstrumentoMapper;
@@ -26,12 +27,16 @@ public class InstrumentoService {
     }
 
     public Instrumento createInstrumento(InstrumentoInput input){
+        if(instrumentoRepository.existsByNome(input.getNome().toUpperCase()))
+            throw new RogueException("Já existe um instrumento com este nome.");
         Instrumento instrumento = new Instrumento();
-        instrumento.setNome(input.getNome());
+        instrumento.setNome(input.getNome().toUpperCase());
         return InstrumentoMapper.entityToDomain(instrumentoRepository.save(InstrumentoMapper.domainToEntity(instrumento)));
     }
 
     public void deleteInstrumento(Long id){
+        if(!instrumentoRepository.existsById(id))
+            throw new RogueException("Não existe um instrumento com este ID.");
         instrumentoRepository.deleteById(id);
     }
 }
